Async execute pipeline job task,avoid blocking zk events
